+++ include/Small/Widgets/Image.hpp
@@ -1,54 +1,53 @@
 /**
- * @file Label.hpp
+ * @file Image.hpp
  * @author Alexandre Plateau (lexplt.dev@gmail.com)
- * @brief Basic label widget
+ * @brief Basic image widget
  * @version 0.1
- * @date 2020-04-18
+ * @date 2020-04-19
  * 
  * @copyright Copyright (c) 2020
  * 
  */
 
-#ifndef sgl_small_widgets_label
-#define sgl_small_widgets_label
+#ifndef sgl_small_widgets_image
+#define sgl_small_widgets_image
 
 #include <Small/Widgets/Base.hpp>
 
-#include <SFML/Graphics/Text.hpp>
-#include <string>
+#include <SFML/Graphics/Sprite.hpp>
 
 namespace sgl::Widgets
 {
     /**
-     * @brief Basic label widget, doesn't listen to events by default
+     * @brief Basic image widget
      * @details The widget must be used inside a sgl::Widgets::Layout to be rendered
      *          properly on screen.
      * 
      */
-    class Label : public Base
+    class Image : public Base
     {
     public:
         /**
-         * @brief Construct a new Label widget
+         * @brief Construct a new Image widget
          * 
          * @param id Unique identifier for the widget (automatically set)
          * @param parent Pointer to the parent widget (automatically set)
          * @param bounds Local bounds of the button
          */
-        Label(int id, Base::Ptr parent, const sf::IntRect& bound);
+        Image(int id, Base::Ptr parent, const sf::IntRect& bounds);
 
         /**
-         * @brief Destroy the Label widget
+         * @brief Destroy the Image widget
          * 
          */
-        ~Label();
+        ~Image();
 
         /**
-         * @brief Return a reference to the sf::Text held, to modify it
+         * @brief Return a reference to the sprite to modify it
          * 
-         * @return sf::Text& 
+         * @return sf::Sprite& 
          */
-        sf::Text& text();
+        sf::Sprite& sprite();
 
         /**
          * @brief Render the widget on screen
@@ -59,7 +58,7 @@ namespace sgl::Widgets
         void onRender(sf::RenderTarget& screen, const sf::Transform& transform);
 
     protected:
-        sf::Text m_text;  //< Text of the label
+        sf::Sprite m_sprite;  //< The image itself
     };
 }